+++ x/fungible/keeper/evm.go
@@ -101,6 +101,12 @@ func (k Keeper) DeployZRC20Contract(
 	if chain == nil {
 		return common.Address{}, cosmoserrors.Wrapf(zetaObserverTypes.ErrSupportedChains, "chain %s not found", chainStr)
 	}
+	// Check if Contract has already been deployed for Asset
+	_, found := k.GetForeignCoinFromAsset(ctx, erc20Contract, chainID)
+	if found {
+		return common.Address{}, types.ErrForeignCoinAlreadyExist
+	}
+
 	system, found := k.GetSystemContract(ctx)
 	if !found {
 		return common.Address{}, cosmoserrors.Wrapf(types.ErrSystemContractNotFound, "system contract not found")
@@ -118,7 +124,6 @@ func (k Keeper) DeployZRC20Contract(
 	if err != nil {
 		return common.Address{}, cosmoserrors.Wrapf(types.ErrABIPack, "failed to deploy ZRC20 contract: %s, %s", name, err.Error())
 	}
-
 	coin, _ := k.GetForeignCoins(ctx, contractAddr.Hex())
 	coin.CoinType = coinType
 	coin.Name = name